chore(bolt-cli): bolt contract bindings
