+++ bolt-cli/src/contracts/bolt.rs
@@ -1,0 +1,20 @@
+use alloy::sol;
+
+sol! {
+    #[allow(missing_docs)]
+    #[sol(rpc)]
+    interface BoltValidators {
+        /// @notice Register a batch of Validators and authorize a Collateral Provider and Operator for them
+        /// @dev This function allows anyone to register a list of Validators.
+        /// @param pubkeyHashes List of BLS public key hashes for the Validators to be registered
+        /// @param maxCommittedGasLimit The maximum gas that the Validator can commit for preconfirmations
+        /// @param authorizedOperator The address of the authorized operator
+        function batchRegisterValidatorsUnsafe(bytes20[] calldata pubkeyHashes, uint32 maxCommittedGasLimit, address authorizedOperator);
+
+        error KeyNotFound();
+        error InvalidQuery();
+        #[derive(Debug)]
+        error ValidatorDoesNotExist(bytes20 pubkeyHash);
+        error InvalidAuthorizedOperator();
+    }
+}
