+++ bolt-cli/src/contracts/mod.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::cli::Chain;
 
+pub mod bolt;
 pub mod eigenlayer;
 pub mod erc20;
